requirements.txt
click>=8.0
PyYAML>=6.0
watchdog>=3.0.0
rich>=13.0
libtmux>=0.16.3
python-dotenv>=1.0
# Optional: packages for AI/tool integrations can be added later
